Remove windows support
I don't use windows, and don't have an easy way to develop locally for
windows. Wepoll doesn't cover all of the features that kqueue and epoll
and will be a blocker if we consider using edge triggered behavior of
epoll and kqueue in the future. I'll consider reviving windows support
again if someone makes a request for it.